Add files via upload
This is the Text Corpus on which the Text Generation model is trained upon. It comprises of the first 5 parts of a story "The Enemy" written by Pearl Sydenstricker Buck. The story revolves around a Japanese surgeon whose name is Sadao. He studies in America, marries a Japanese girl named Hana there and comes back to Japan to settle down. World War II broke during that time. All the doctors were under the obligation to go to the Japanese army. However, Sadao stayed back with an old general. As the old general was ailing, he was in need of Sadao. However, an American Navy man finds his way into Sadao’s life. As the soldier was not well, Sadao offers him medical help. Sadao was not willing to help the enemy but he does that anyway. He knows the risk of helping the enemy. Due to this, Sadao conspires to kill the soldier in his sleep. However, Sadao decides not to do so because of humanity.